+++ src/policy.rs
@@ -28,6 +28,12 @@ impl BitOr for EncryptionHint {
     }
 }
 
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AxisAttributePorperties {
+    pub name: String,
+    pub encryption_hint: EncryptionHint,
+}
+
 /// Defines a policy axis by its name and its underlying attribute properties.
 /// An attribute property defines its name and a hint about whether hybridized
 /// encryption should be used for it (hint set to `true` if this is the case).
@@ -39,7 +45,7 @@ pub struct PolicyAxis {
     /// Axis name
     pub name: String,
     /// Names of the axis attributes and hybridized encryption hints
-    pub attribute_properties: Vec<(String, EncryptionHint)>,
+    pub attributes_properties: Vec<AxisAttributePorperties>,
     /// `true` if the axis is hierarchical
     pub hierarchical: bool,
 }
@@ -54,14 +60,17 @@ impl PolicyAxis {
     #[must_use]
     pub fn new(
         name: &str,
-        attribute_properties: Vec<(&str, EncryptionHint)>,
+        attributes_properties: Vec<(&str, EncryptionHint)>,
         hierarchical: bool,
     ) -> Self {
         Self {
             name: name.to_string(),
-            attribute_properties: attribute_properties
+            attributes_properties: attributes_properties
                 .into_iter()
-                .map(|(axis_name, hint)| (axis_name.to_string(), hint))
+                .map(|(axis_name, encryption_hint)| AxisAttributePorperties {
+                    name: axis_name.to_string(),
+                    encryption_hint,
+                })
                 .collect(),
             hierarchical,
         }
@@ -70,20 +79,30 @@ impl PolicyAxis {
     /// Returns the number of attributes belonging to this axis.
     #[must_use]
     pub fn len(&self) -> usize {
-        self.attribute_properties.len()
+        self.attributes_properties.len()
     }
 
     /// Return `true` if the attribute list is empty
     #[must_use]
     pub fn is_empty(&self) -> bool {
-        self.attribute_properties.is_empty()
+        self.attributes_properties.is_empty()
     }
 }
 
-/// A policy is a set of policy axes. A fixed number of attribute creations
-/// (revocations + additions) is allowed.
 #[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct Policy {
+pub struct PolicyAxesParameters {
+    pub attribute_names: Vec<String>,
+    pub is_hierarchical: bool,
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct PolicyAttributesParameters {
+    pub values: Vec<u32>,
+    pub encryption_hint: EncryptionHint,
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct LegacyPolicy {
     /// Last value taken by the attriute.
     pub(crate) last_attribute_value: u32,
     /// Maximum attribute value. Defines a maximum number of attribute
@@ -91,9 +110,32 @@ pub struct Policy {
     pub max_attribute_creations: u32,
     /// Policy axes: maps axes name to the list of associated attribute names
     /// and a boolean defining whether or not this axis is hierarchical.
-    pub axes: HashMap<String, (Vec<String>, bool)>,
+    pub axes: HashMap<String, PolicyAxesParameters>,
     /// Maps an attribute to its values and its hybridization hint.
-    pub attributes: HashMap<Attribute, (Vec<u32>, EncryptionHint)>,
+    pub attributes: HashMap<Attribute, Vec<u32>>,
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum PolicyVersion {
+    V1,
+}
+
+/// A policy is a set of policy axes. A fixed number of attribute creations
+/// (revocations + additions) is allowed.
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct Policy {
+    /// Version number
+    pub version: PolicyVersion,
+    /// Last value taken by the attriute.
+    pub(crate) last_attribute_value: u32,
+    /// Maximum attribute value. Defines a maximum number of attribute
+    /// creations (revocations + addition).
+    pub max_attribute_creations: u32,
+    /// Policy axes: maps axes name to the list of associated attribute names
+    /// and a boolean defining whether or not this axis is hierarchical.
+    pub axes: HashMap<String, PolicyAxesParameters>,
+    /// Maps an attribute to its values and its hybridization hint.
+    pub attributes: HashMap<Attribute, PolicyAttributesParameters>,
 }
 
 impl Display for Policy {
@@ -107,11 +149,49 @@ impl Display for Policy {
 }
 
 impl Policy {
+    /// Converts the given string into a Policy. Does not fail if the given
+    /// string uses the legacy format.
+    pub fn parse_and_convert(bytes: &[u8]) -> Result<Self, Error> {
+        match serde_json::from_slice(bytes) {
+            Ok(policy) => Ok(policy),
+            Err(e) => {
+                if let Ok(policy) = serde_json::from_slice::<LegacyPolicy>(bytes) {
+                    // Convert the legacy format to the current one.
+                    Ok(Policy {
+                        version: PolicyVersion::V1,
+                        max_attribute_creations: policy.max_attribute_creations,
+                        last_attribute_value: policy.last_attribute_value,
+                        axes: policy.axes,
+                        attributes: policy
+                            .attributes
+                            .into_iter()
+                            .map(|(name, values)| {
+                                (
+                                    name,
+                                    PolicyAttributesParameters {
+                                        values,
+                                        encryption_hint: EncryptionHint::Classic,
+                                    },
+                                )
+                            })
+                            .collect(),
+                    })
+                } else {
+                    // Return the `Policy` deserialization error message instead of the
+                    // `LegacyPolicy` one since this is the one that should be used.
+                    Err(Error::DeserializationError(e))
+                }
+            }
+        }
+    }
+
     /// Generates a new policy object with the given number of attribute
     /// creation (revocation + addition) allowed.
+    #[inline]
     #[must_use]
     pub fn new(nb_creations: u32) -> Self {
         Self {
+            version: PolicyVersion::V1,
             last_attribute_value: 0,
             max_attribute_creations: nb_creations,
             axes: HashMap::new(),
@@ -120,7 +200,7 @@ impl Policy {
     }
 
     /// Returns the remaining number of allowed attribute creations (additions + rotations).
-    #[must_use]
+    #[inline]
     pub fn remaining_attribute_creations(&self) -> u32 {
         self.max_attribute_creations - self.last_attribute_value
     }
@@ -138,23 +218,32 @@ impl Policy {
         if self.axes.get(&axis.name).is_some() {
             return Err(Error::ExistingPolicy(axis.name));
         }
-        let mut axis_attributes = Vec::with_capacity(axis.attribute_properties.len());
+        let mut axis_attributes = Vec::with_capacity(axis.attributes_properties.len());
 
-        for (attr_name, is_hybridized) in axis.attribute_properties {
+        for properties in axis.attributes_properties {
             self.last_attribute_value += 1;
-            axis_attributes.push(attr_name.clone());
-            let attribute = (axis.name.clone(), attr_name).into();
+            axis_attributes.push(properties.name.clone());
+            let attribute = (axis.name.clone(), properties.name.clone()).into();
             if self.attributes.get(&attribute).is_some() {
                 return Err(Error::ExistingPolicy(format!("{attribute:?}")));
             }
             self.attributes.insert(
                 attribute,
-                ([self.last_attribute_value].into(), is_hybridized),
+                PolicyAttributesParameters {
+                    values: [self.last_attribute_value].into(),
+                    encryption_hint: properties.encryption_hint,
+                },
             );
         }
 
-        self.axes
-            .insert(axis.name, (axis_attributes, axis.hierarchical));
+        self.axes.insert(
+            axis.name,
+            PolicyAxesParameters {
+                attribute_names: axis_attributes,
+                is_hierarchical: axis.hierarchical,
+            },
+        );
+
         Ok(())
     }
 
@@ -163,9 +252,9 @@ impl Policy {
     pub fn rotate(&mut self, attr: &Attribute) -> Result<(), Error> {
         if self.last_attribute_value == self.max_attribute_creations {
             Err(Error::CapacityOverflow)
-        } else if let Some((heap, _)) = self.attributes.get_mut(attr) {
+        } else if let Some(attribute_parameters) = self.attributes.get_mut(attr) {
             self.last_attribute_value += 1;
-            heap.push(self.last_attribute_value);
+            attribute_parameters.values.push(self.last_attribute_value);
             Ok(())
         } else {
             Err(Error::AttributeNotFound(attr.to_string()))
@@ -173,6 +262,7 @@ impl Policy {
     }
 
     /// Returns the list of Attributes of this Policy.
+    #[inline]
     #[must_use]
     pub fn attributes(&self) -> Vec<Attribute> {
         self.attributes.keys().cloned().collect::<Vec<Attribute>>()
@@ -180,29 +270,34 @@ impl Policy {
 
     /// Returns the list of all values given to this attribute over rotations.
     /// The current value is returned first
+    #[inline]
     pub fn attribute_values(&self, attribute: &Attribute) -> Result<Vec<u32>, Error> {
         self.attributes
             .get(attribute)
-            .map(|(values, _)| values.iter().rev().copied().collect())
+            .map(|attribute_parameters| attribute_parameters.values.iter().rev().copied().collect())
             .ok_or_else(|| Error::AttributeNotFound(attribute.to_string()))
     }
 
     /// Returns the hybridization hint of the given attribute.
+    #[inline]
     pub fn attribute_hybridization_hint(
         &self,
         attribute: &Attribute,
     ) -> Result<EncryptionHint, Error> {
         self.attributes
             .get(attribute)
-            .map(|(_, is_hybridized)| *is_hybridized)
+            .map(|attribute_parameters| attribute_parameters.encryption_hint)
             .ok_or_else(|| Error::AttributeNotFound(attribute.to_string()))
     }
 
     /// Retrieves the current value of an attribute.
+    #[inline]
     pub fn attribute_current_value(&self, attribute: &Attribute) -> Result<u32, Error> {
         self.attributes
             .get(attribute)
-            .map(|(values, _)| values[values.len() - 1])
+            .map(|attribute_parameters| {
+                attribute_parameters.values[attribute_parameters.values.len() - 1]
+            })
             .ok_or_else(|| Error::AttributeNotFound(attribute.to_string()))
     }
 }